Update tox.ini to use py37,py38 instead of ci37,ci38
As defined in the tox documentation
(https://tox.readthedocs.io/en/latest/config.html#tox-environments),
environment with names "pyNM" have special meaning and implicitly define
the Python interpreter.

Environment names ci37 and ci38 don't have any special meaning and run
the default version of the interpreter. This results in that that both
environments use the same interpreter locally.

The PR changes the names of environments to make sure that tox creates a
virtual environment with the correct interpreter version.